size

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	float floatType;
 
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(charType));
-	printf("Size of an int: %zu byte(s)\n", sizeof(intType));
-	printf("Size of a long int: %zu byte(s)\n",  sizeof(longType));
-	printf("Size of a long long int: %zu byte(s)\n", sizeof(longerType));
-	printf("Size of an float: %zu byte(s)\n", sizeof(floatType));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(charType));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(intType));
+	printf("Size of a long int: %lu byte(s)\n",  (unsigned long)sizeof(longType));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(longerType));
+	printf("Size of an float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
 
 	return (0);
